+++ test/integration/helpers/config.js
@@ -37,7 +37,12 @@ const defaultIntegrationConfig = {
     localSaveNodeAdminMaxLines: 800,
     localSaveNoteMaxLines: 20,
     ideEditOpenParent: false,
-    ideVmsNoteEnabled: false
+    ideVmsNoteEnabled: false,
+    ideObjectBrowserEnabled: true,
+    idePropertyBrowserEnabled: true,
+    ideHoverOverlaysEnabled: true,
+    ideReferenceNavigationEnabled: true,
+    ideScratchEnabled: true
   },
   shorten: {
     enabled: false,